fix 8ef9bb4 mistakes
Change to BigInteger from Integer.

+++ zerotwobot/modules/sql/users_sql.py
@@ -50,7 +50,7 @@ class ChatMembers(BASE):
         nullable=False,
     )
     user = Column(
-        Integer,
+        BigInteger,
         ForeignKey("users.user_id", onupdate="CASCADE", ondelete="CASCADE"),
         nullable=False,
     )
